move b_coil_axis
