+++ bluemira/magnetostatics/finite_element_2d.py
@@ -31,6 +31,17 @@ import numpy as np
 from bluemira.base.constants import MU_0
 
 
+def b_coil_axis(r, z, pz, curr):
+    """
+    Return the module of the magnetic field of a coil (of radius r and centered in
+    (0, z)) calculated on a point on the coil axis at a distance pz from the
+    axis origin.
+
+    TODO: add equation
+    """
+    return 4 * np.pi * 1e-7 * curr * r**2 / (r**2 + (pz - z) ** 2) ** 1.5 / 2.0
+
+
 def _convert_const_to_dolfin(value: float):
     """Convert a constant value to a dolfin function"""
     if not isinstance(value, (int, float)):
