+++ examples/equilibria/fem_fixed_boundary/fem_magnetostatic_2D_single_coil.ex.py
@@ -62,11 +62,12 @@ import bluemira.geometry.tools as tools
 import bluemira.magnetostatics.greens as greens
 from bluemira.base.components import Component, PhysicalComponent
 from bluemira.base.file import get_bluemira_path
-from bluemira.equilibria.fem_fixed_boundary.fem_magnetostatic_2D import (
-    FemMagnetostatic2d,
-)
-from bluemira.equilibria.fem_fixed_boundary.utilities import ScalarSubFunc, b_coil_axis
 from bluemira.geometry.face import BluemiraFace
+from bluemira.magnetostatics.finite_element_2d import (
+    FemMagnetostatic2d,
+    ScalarSubFunc,
+    b_coil_axis,
+)
 from bluemira.mesh import meshing
 from bluemira.mesh.tools import import_mesh, msh_to_xdmf
